+++ xiaozhi-ai/src/main/java/com/xiaozhi/ai/tts/providers/VolcengineTtsService.java
@@ -39,7 +39,7 @@ public class VolcengineTtsService implements TtsService {
 
     private final OkHttpClient client = HttpUtil.client;
 
-    public VolcengineTtsService(ConfigBO config, String voiceName, Float pitch, Float speed, String outputPath) {
+    public VolcengineTtsService(ConfigBO config, String voiceName, Double pitch, Double speed, String outputPath) {
         this.options = XiaozhiTtsOptions.builder().voiceName(voiceName).pitch(pitch).speed(speed).build();
         this.outputPath = outputPath;
         this.appId = config.getAppId();
