+++ xiaozhi-common/src/main/java/com/xiaozhi/common/model/resp/RoleResp.java
@@ -26,10 +26,10 @@ public class RoleResp {
     private String voiceName;
 
     @Schema(description = "语音音调")
-    private Float ttsPitch;
+    private Double ttsPitch;
 
     @Schema(description = "语音语速")
-    private Float ttsSpeed;
+    private Double ttsSpeed;
 
     @Schema(description = "状态(1启用 0禁用)")
     private String state;
